% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{load.save}
\alias{load.save}
\title{Auxilary function for loading and saving files}
\usage{
load.save(filename)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Auxilary function for loading and saving files
}
\examples{
load.save()
}
\keyword{cats}
