% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{SRSFN_wrapper}
\alias{SRSFN_wrapper}
\title{A script wrapper}
\usage{
SRSFN_wrapper(filenames, TSSregions, MARG = 0, CUTOFF = 100, ADJ = 0.5,
  p = 1000, weight.adj = 0, penalty = 1e-04, quant = 0.999,
  max.cores = detectCores())
}
\arguments{
\item{filenames}{names of the UCSC bed files containing read coordinates}

\item{TSSregions}{name of the UCSC bed file specifying genomic regions for the shape analysis (e.g. +/- 1000bp intervals from TSS)}

\item{MARG}{Default is 0. Specifies by how many bp should the regions (specified in TSSregions) be expended by. Allows negative numbers}

\item{CUTOFF}{Default is 100. Specifies minimum number of reads per region of interest for the region to be included in the analysis. Low read counts per region could result in unreliable and noisy read density shapes.}

\item{ADJ}{Default is 0.5. Bandwidth adjustment for density estimation}

\item{p}{Default is 1000. Size of the margins of the left and right end of genomic region to account for unnecessary shifts in empty regions (where read density is zero). This is different then MARG parameter and is only used to clear possible algorithm errors.}

\item{weight.adj}{Default is 0. Weight of the read coordinates in estimating the read desnity over region of interest.}

\item{penalty}{Default is 0.0001. Controls the penalty for the amount of shift.}

\item{quant}{Default is 0.999 Top quantile of the regions with the highest shift statistic that are to be visualized}

\item{max.cores}{Default is given by detectCores() function.}
}
\description{
This function wraps toghether the script to perform the SRSF shape analysis of reads on specified genomic regions
}
\examples{
TSSregions.bed.name="data/chr1.TSSgencode.bed";
MARG = 0; CUTOFF=100;  ADJ=0.5;p=1000;weight.adj=0;penalty=0.0001; quant=0.999; max.cores=56;
group.id<-NULL #Need to adjust for variable group sizes
filenames=c("data/chr1.GH1_cap_R1_clip.fastq_q20.bed", "data/chr1.GH2_cap_R1_clip.fastq_q20.bed",
            "data/chr1.IH1_cap_R1_clip.fastq_q20.bed", "data/chr1.IH2_cap_R1_clip.fastq_q20.bed")
results<-SRSFN_wrapper(filenames,TSSregions)
}
\keyword{main}
\keyword{wrapper}
